1. 의존성 관련 파일 추가 & 10. primary key
