+++ models/__init__.py
@@ -1,4 +1,7 @@
 from peewee import SqliteDatabase
+import sqlite3
 
+connection = sqlite3.connect('test_sqlite.db')
+curser = connection.cursor()
 
 DB = SqliteDatabase('test_peewee.db')
