--- requirements.txt
@@ -1,6 +1,0 @@
-Flask
-flask-restful
-flask-bcrypt
-flask-httpauth
-peewee
-uwsgi
