+++ run.py
@@ -1,5 +1,4 @@
 from models.Database import initialize
-from app import app
 
 
 initialize()
